chenged event added endVoting check more tests
