+++ Error description.txt
@@ -20,7 +20,7 @@ Error codes
 17: Case does not exist. -> MultiSig.sol
 18: This address has already approved this case. -> MultiSig.sol
 19: Case open for voting, cant be approved. -> MultiSig.sol
-20: Alternatives cannot be added because the case is open for voting or deactivated. -> Case.sol
+20: Alternatives cannot be added because the case: does not exist OR is open for voting OR is deactivated. -> Case.sol
 21: OnlyOwners required. -> MultiSig.sol
 22: OnlyOwners required. -> Case.sol
 23: OnlyOwners required. -> Case.sol
@@ -36,6 +36,7 @@ Error codes
 33: You have moved after the case was created. -> Case.sol
 34: Number of SuperAdmin has to be an odd number to avoid a tie. -> Case.sol
 35: Nunber of SuperAdmin has to be an odd number. -> Proxy.sol
-36: 
+36: Case does not exist. -> Case.sol
+37:
