+++ FrontEnd/abi.js
@@ -2,7 +2,7 @@ window.abi = getAbi;
 
 function getAbi() {
     return new Promise((res) => {
-        $.getJSON("../build/contracts/Token.json", ((json) => {
+        $.getJSON("../BackEnd/build/contracts/Token.json", ((json) => {
             res(json.abi);
         }))
     })
